refactor(library/Attempt): moves "Attempt/Outcome/Success/exports.ts" to "Attempt/Outcome/Success/definition.ts"
