+++ src/library/Attempt/Outcome/Success/definition.ts
